requirements.txt
# Core data science libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
nbconvert>=7.0.0

# Data visualization
plotly>=5.14.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.2.0

# Data validation
great-expectations>=0.17.0

# Environment management
python-dotenv>=1.0.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Excel support
openpyxl>=3.1.0
